src/strategies/statistical_arbitrage/strategy.py
from typing import Dict, List
from decimal import Decimal
import numpy as np
from datetime import datetime

class StatisticalArbitrageStrategy:
    """统计套利策略类
    
    通过统计分析识别价格偏离并进行套利交易。
    """
    
    def __init__(self, config: Dict = None):
        """初始化统计套利策略
        
        Args:
            config: 策略配置，若未提供则使用默认配置
        """
        self.config = config or DEFAULT_CONFIG.copy()
        validate_config(self.config)
        
        # 初始化状态
        self.symbol = self.config['symbol']
        self.window_size = self.config['window_size']
        self.z_threshold = self.config['z_threshold']
        self.price_history = []
        self.active_orders: Dict[str, Dict] = {}
        self.positions: Dict[str, Decimal] = {}
        
    def _calculate_zscore(self, price: Decimal) -> float:
        """计算当前价格的Z分数
        
        Args:
            price: 当前价格
            
        Returns:
            float: Z分数
        """
        if len(self.price_history) < self.window_size:
            return 0.0
            
        prices = np.array(self.price_history[-self.window_size:])
        mean = np.mean(prices)
        std = np.std(prices)
        
        if std == 0:
            return 0.0
            
        return (float(price) - mean) / std
    
    def _execute_trades(self, zscore: float, current_price: Decimal):
        """执行套利交易
        
        Args:
            zscore: Z分数
            current_price: 当前价格
        """
        amount = Decimal(str(self.config['trade_amount']))
        
        if abs(zscore) > self.z_threshold:
            if zscore > 0:  # 价格高于均值
                self._place_order('sell', amount, current_price)
            else:  # 价格低于均值
                self._place_order('buy', amount, current_price)
    
    def _place_order(self, side: str, amount: Decimal, price: Decimal):
        """下单
        
        Args:
            side: 交易方向
            amount: 交易数量
            price: 交易价格
        """
        # 实现下单逻辑
        pass
    
    def on_tick(self, tick_data: Dict):
        """处理市场数据更新
        
        Args:
            tick_data: 市场数据字典
        """
        current_price = Decimal(str(tick_data['price']))
        self.price_history.append(float(current_price))
        
        # 保持价格历史在窗口大小范围内
        if len(self.price_history) > self.window_size * 2:
            self.price_history = self.price_history[-self.window_size * 2:]
        
        # 计算Z分数并执行交易
        if len(self.price_history) >= self.window_size:
            zscore = self._calculate_zscore(current_price)
            if self._validate_opportunity(zscore):
                self._execute_trades(zscore, current_price)
    
    def _validate_opportunity(self, zscore: float) -> bool:
        """验证套利机会是否可行
        
        Args:
            zscore: Z分数
            
        Returns:
            bool: 是否可行
        """
        # 检查是否超过阈值且没有达到最大持仓
        if abs(zscore) <= self.z_threshold:
            return False
            
        total_position = sum(abs(pos) for pos in self.positions.values())
        if total_position >= self.config['max_position']:
            return False
            
        return True
    
    def on_trade(self, trade: Dict):
        """处理成交信息
        
        Args:
            trade: 成交信息
        """
        # 更新持仓信息
        self._update_position(trade)
    
    def _update_position(self, trade: Dict):
        """更新持仓信息
        
        Args:
            trade: 成交信息
        """
        symbol = trade['symbol']
        amount = Decimal(str(trade['amount']))
        
        if trade['side'] == 'buy':
            self.positions[symbol] = self.positions.get(symbol, Decimal('0')) + amount
        else:
            self.positions[symbol] = self.positions.get(symbol, Decimal('0')) - amount
    
    def start(self):
        """启动策略"""
        # 实现策略启动逻辑
        pass
    
    def stop(self):
        """停止策略"""
        # 实现策略停止逻辑
        pass