requirements.txt
# 基础依赖包
# 这些包是项目运行所必需的基础组件
setuptools==75.8.0
wheel==0.45.1
typing_extensions==4.12.2
pydantic==2.10.1
pydantic_core==2.27.1
annotated-types==0.7.0

# 网络通信相关
# 用于处理HTTP请求、WebSocket连接和网络通信的包
aiohttp==3.10.11
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiosignal==1.3.2
requests==2.32.3
uvicorn==0.24.0
fastapi==0.104.1
starlette==0.41.3
yarl==1.18.3
multidict==6.1.0
frozenlist==1.5.0
pycares==4.5.0
websockets==12.0

# 数据库相关
# 用于数据库连接和ORM操作的包
aiomysql==0.2.0
aiosqlite==0.20.0
PyMySQL==1.1.1
tortoise-orm==0.22.1
pypika-tortoise==0.3.0

# 数据处理和缓存
# 用于数据处理、缓存管理和时间处理的包
propcache==0.2.1
iso8601==2.1.0
pytz==2024.2

# 加密和安全
# 用于处理加密、证书和安全相关的包
cryptography==44.0.0
certifi==2025.1.31
idna==3.10
charset-normalizer==3.4.1
cffi==1.17.1
pycparser==2.22

# 交易所API
# 用于与加密货币交易所交互的包
ccxt==4.4.52

# 工具包
# 其他工具和辅助功能包
attrs==25.1.0
urllib3==2.3.0
python-dotenv==1.0.1
psutil==5.9.6
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
